fix: ensure resources ignition directory exists (#2684)

+++ yarn-project/circuits.js/src/crs/index.ts
@@ -3,7 +3,7 @@ import { fileURLToPath } from '@aztec/foundation/url';
 
 import isNode from 'detect-node';
 import { existsSync } from 'fs';
-import { open } from 'fs/promises';
+import { mkdir, open } from 'fs/promises';
 import { dirname, join } from 'path';
 
 /**
@@ -180,8 +180,12 @@ export class Crs {
     await this.crs.init();
     if (isNode) {
       const localPath = dirname(fileURLToPath(import.meta.url)) + this.localPath;
-      // save downloaded CRS on file
-      if (this.saveOnFile && !existsSync(localPath)) {
+      const devPath = join(fileURLToPath(import.meta.url), this.devPath);
+      // save downloaded CRS on file if we don't have a local or dev version
+      if (this.saveOnFile && !existsSync(localPath) && !existsSync(devPath)) {
+        const dir = localPath.substring(0, localPath.lastIndexOf('/'));
+        await mkdir(dir, { recursive: true });
+
         const fileHandle = await open(localPath, 'w');
         const g1Data = Buffer.from(this.crs.getG1Data());
         try {
